pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "va-shared"
version = "0.1.23"
description = "Shared utilities for Ensi AI Voice Assistant"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "pydantic>=2.0.0",
    "firebase-admin>=6.0.0",
    "aiohttp>=3.8.0",
    "ulid-py>=1.1.0",
]

[project.scripts]
bump-version = "scripts.bump_version:main"
